move map_api and line shared code
