+++ apps/site/test/site_web/controllers/schedule/line_test.exs
@@ -1,6 +1,7 @@
 defmodule SiteWeb.ScheduleController.LineTest do
   use SiteWeb.ConnCase, async: true
   import SiteWeb.ScheduleController.Line
+  import SiteWeb.ScheduleController.Line.Helpers
   alias Stops.{RouteStop, RouteStops}
 
   doctest SiteWeb.ScheduleController.Line
